APIproject/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-filter==25.1
drf-yasg==1.21.10
pytz==2025.2
sqlparse==0.5.3